+++ src/playsync.rs
@@ -85,7 +85,7 @@ mod test {
     #[test]
     fn test_cancel_handle_cancelled() {
         let cancel_handle = CancelHandle::new();
-        assert_eq!(false, cancel_handle.is_cancelled());
+        assert!(!cancel_handle.is_cancelled());
 
         let join = {
             let cancel_handle = cancel_handle.clone();
@@ -94,13 +94,13 @@ mod test {
 
         cancel_handle.cancel();
         assert!(join.join().is_ok());
-        assert_eq!(true, cancel_handle.is_cancelled());
+        assert!(cancel_handle.is_cancelled());
     }
 
     #[test]
     fn test_cancel_handle_expired() {
         let cancel_handle = CancelHandle::new();
-        assert_eq!(false, cancel_handle.is_cancelled());
+        assert!(!cancel_handle.is_cancelled());
 
         let join = {
             let cancel_handle = cancel_handle.clone();
@@ -109,6 +109,6 @@ mod test {
 
         cancel_handle.expire();
         assert!(join.join().is_ok());
-        assert_eq!(false, cancel_handle.is_cancelled());
+        assert!(!cancel_handle.is_cancelled());
     }
 }