+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 DMX engine using OLA has been added. Contains a built in dimming engine.
 
+Added the ability to exclude MIDI channels from MIDI playback.
+
 ## [0.1.5] - Initial DMX engine, fixed MIDI clock timing.
 
 Initial DMX engine implemented. This is not quite ready for prime time.